refactor(okf_panel): extract helpers and tree modules
Split the panel's boot IIFE so its sidebar logic is reachable from unit
tests. helpers.js holds the pure DOM-leaf builders; tree.js owns the sidebar
tree and the highlight, receiving its DOM handles and a concept-select
callback through initTree() so it never imports the entry.

Drop the entry's @ts-nocheck and resolve the resulting diagnostics: a local
requireEl() narrows the five required element handles once at their
declaration instead of at each use, the search input is typed as
HTMLInputElement, and the remaining parameters gain JSDoc annotations.
Unused caught-error bindings become optional catch bindings, which lets the
entry leave the no-unused-vars exemption list.
